+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Deployment set - Introducing deployment set (experimental feature)
 - CI - concurrency and build on tag push
 
 ## [6.0.6] - 2022-11-17